+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 All notable changes to this project will be documented in this file.
-"Unreleased" changes are implemented but haven't been uploaded yet.
+"Unreleased" changes are implemented but haven't been released yet.
 
 ## [2.0.2] - 2018-03-28
   - Fixed Ctrl vs Cmd issue with shortcut tray.